% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_Data_Documentation.R
\docType{data}
\name{BayesDat}
\alias{BayesDat}
\title{Sample Data for Analysis}
\format{
This is a list
\describe{
\item{var1}{this is a description}
\item{var2}{this is another descritipn}
}
}
\usage{
BayesDat
}
\description{
this is some data
}
\keyword{datasets}
