% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes_MSM.R
\name{BayesMSM}
\alias{BayesMSM}
\title{A BayesMSM Function}
\usage{
BayesMSM(
  Data,
  PatID,
  Treat_Vars,
  Outcome_Var,
  Treat_M_List,
  Outcome_M_List,
  MSM_Model = NULL,
  G_List = NULL,
  Psi,
  DR = FALSE,
  Bayes = FALSE,
  Normalized = FALSE,
  B = 1,
  Bayes_Seed = 1
)
}
\arguments{
\item{Data}{dataset used for analysis}

\item{PatID}{patient identified}

\item{Outcome_Var}{outcome variable}

\item{Treat_M_List}{list of treatment models}

\item{Outcome_M_List}{list of outcome models}

\item{MSM_Model}{model for continuous MSM}

\item{G_List}{decision rule list}

\item{Psi}{points to explore in the parameter space}

\item{DR}{boolean indicating whether double robust estimator should be used}

\item{Bayes}{boolean indicating whether Bayesian methods will be used}

\item{Normalized}{boolean indicating whether to use normalized or non-normalized estimators}

\item{B}{number of Bayesian bootstrap samples}

\item{Bayes_Seed}{seed to control random number generation}

\item{Treat_vars}{treatment variables}
}
\description{
This function allows you to fit a Bayesian MSM
}
\keyword{Bayes}
