% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_Optim_Export.R
\name{FitInfer}
\alias{FitInfer}
\title{Credible Intervals for Optimizing Regime}
\usage{
FitInfer(
  Design_Object,
  Boot_Start,
  Boot_End,
  Psi_new,
  N,
  Location = NA,
  Additional_Samp,
  Control_Genoud = list()
)
}
\arguments{
\item{Psi_new}{grid of points where posterior paths will be evaluated}

\item{N}{number of posterior samples}

\item{Location}{Location where to save results for bootstrapped samples, in case process is interrupted}

\item{Control_Genoud}{parameters to be fed into the genoud function}

\item{Sequential_Object}{object returned from SequenceFit}

\item{Boot}{Number of Bayesian bootstrap samples}
}
\value{
this function returns a matrix with the full posterior draws
}
\description{
This function allows the user to obtain credible intervals for the optimizing regimes
}
