% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_Optim_Export.R
\name{PostMean}
\alias{PostMean}
\title{Obtaining posterior mean of Gaussian process.}
\usage{
PostMean(x, GP_Object)
}
\arguments{
\item{GP_Object}{an object returned from either DesignFit or SequenceFit}
}
\description{
function fits the posterior mean.
}
