% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_Optim_Export.R
\name{SequenceFit}
\alias{SequenceFit}
\title{Fitting Gaussian Process parameters for sequential samples}
\usage{
SequenceFit(Previous_Fit, Additional_Samp, Control_Genoud = list(), Psi_new, N)
}
\arguments{
\item{Previous_Fit}{an object that has resulted from DesignFit or SequenceFit}

\item{Additional_Samp}{Number of points to sample sequentially}

\item{Control_Genoud}{}

\item{Bayes_Seed}{allows user to set seed}
}
\value{
This function returns the input parameters required for further fitting of the Gaussian Process when more samples are added
        as well as the current values of the parameters
}
\description{
This function fits Gaussian Process parameters to an initial set of design points in addition to those sampled via the Expected
             improvement criterion
}
